add situation
